requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
requests>=2.26.0 
pexpect>=4.9.0
nuclei>=2.9.0
sqlmap>=1.7.0
sslyze>=5.0.0
python-nmap>=0.7.1
beautifulsoup4>=4.12.0
lxml>=4.9.0